A/requirements.txt
Django>=3.1.2
django-cors-headers==3.4.0
django-extensions==3.0.3
djangorestframework==3.11.0
Markdown==3.2.2
pydot==1.4.1
pyparsing==2.4.7
pytz==2020.1
six==1.15.0
sqlparse==0.3.1